requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
firebase-admin
python-dotenv
